+++ previous-versions/accounting/RELEASES.md
@@ -689,3 +689,13 @@ Based on:
 - [typescript v0.38.0] previous-versions/accounting
 ### Releases
 - [NPM v0.38.0] https://www.npmjs.com/package/@codat/accounting/v/0.38.0 - previous-versions/accounting
+
+## 2023-11-06 01:25:34
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Accounting.yaml
+- Speakeasy CLI 1.112.1 (2.173.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.39.0] previous-versions/accounting
+### Releases
+- [NPM v0.39.0] https://www.npmjs.com/package/@codat/accounting/v/0.39.0 - previous-versions/accounting